+++ ITBees.DragArtboard/Interfaces/Models/ArtboardElementUm.cs
@@ -6,9 +6,9 @@ public class ArtboardElementUm
 {
     public Guid Guid { get; set; }
     public string CustomSerializedObject { get; set; }
-    public int LocationX { get; set; }
-    public int LocationY { get; set; }
-    public int LocationZ { get; set; }
+    public float LocationX { get; set; }
+    public float LocationY { get; set; }
+    public float LocationZ { get; set; }
     public int ZIndex { get; set; }
     public int Width { get; set; }
     public int Height { get; set; }
